update file structure for js classes and tests, update tests for card to run with Mocha
