+++ .gitignore
@@ -1,1 +1,2 @@
 .vscode/
+node_modules/
