+++ test/Trick.test.js
@@ -1,0 +1,10 @@
+// import assert
+import assert from 'assert';
+import { Trick } from '../scripts/classes.mjs';
+
+// Naive approach
+describe('trick', () => {
+  describe('valid trick', () => {
+    
+  });
+});
